Correct behavior when no action is performed
When rendering the grid field *without* calling handleAction() (e.g.
when entering the page the first time or with hitting the back button)
the paginator was always enabled. This is wrong: the user would expect
to see the previous state when using the back button.

Corrected the behavior by leveraging the GridField_DataManipulator
interface. In this case the component must be added before the
GridFieldPaginator component to be able to modify the its behavior.